pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "svdownloader"
version = "0.1.0"
description = "Download Google Street View panoramas at native quality"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "Pillow>=10.0",
    "tqdm>=4.60",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
svdownload = "svdownloader.cli:main"

[tool.setuptools.packages.find]
include = ["svdownloader*"]
